op_ref: Move tests out of eq_op file
