+++ tests/ui/eq_op.rs
@@ -3,6 +3,7 @@
 #[allow(clippy::identity_op, clippy::double_parens, clippy::many_single_char_names)]
 #[allow(clippy::no_effect, unused_variables, clippy::unnecessary_operation, clippy::short_circuit_statement)]
 #[warn(clippy::nonminimal_bool)]
+#[allow(unused)]
 fn main() {
     // simple values and comparisons
     1 == 1;
@@ -49,42 +50,6 @@ fn main() {
     a == a;
     2*a.len() == 2*a.len(); // ok, functions
     a.pop() == a.pop(); // ok, functions
-
-    use std::ops::BitAnd;
-    struct X(i32);
-    impl BitAnd for X {
-        type Output = X;
-        fn bitand(self, rhs: X) -> X {
-            X(self.0 & rhs.0)
-        }
-    }
-    impl<'a> BitAnd<&'a X> for X {
-        type Output = X;
-        fn bitand(self, rhs: &'a X) -> X {
-            X(self.0 & rhs.0)
-        }
-    }
-    let x = X(1);
-    let y = X(2);
-    let z = x & &y;
-
-    #[derive(Copy, Clone)]
-    struct Y(i32);
-    impl BitAnd for Y {
-        type Output = Y;
-        fn bitand(self, rhs: Y) -> Y {
-            Y(self.0 & rhs.0)
-        }
-    }
-    impl<'a> BitAnd<&'a Y> for Y {
-        type Output = Y;
-        fn bitand(self, rhs: &'a Y) -> Y {
-            Y(self.0 & rhs.0)
-        }
-    }
-    let x = Y(1);
-    let y = Y(2);
-    let z = x & &y;
 
     check_ignore_macro();
