+++ tests/ui/op_ref.stderr
@@ -1,5 +1,5 @@
 error: needlessly taken reference of both operands
-  --> $DIR/op_ref.rs:10:15
+  --> $DIR/op_ref.rs:12:15
    |
 LL |     let foo = &5 - &6;
    |               ^^^^^^^
@@ -11,12 +11,20 @@ LL |     let foo = 5 - 6;
    |               ^   ^
 
 error: taken reference of right operand
-  --> $DIR/op_ref.rs:18:8
+  --> $DIR/op_ref.rs:20:8
    |
 LL |     if b < &a {
    |        ^^^^--
    |            |
    |            help: use the right value directly: `a`
 
-error: aborting due to 2 previous errors
+error: taken reference of right operand
+  --> $DIR/op_ref.rs:57:13
+   |
+LL |     let z = x & &y;
+   |             ^^^^--
+   |                 |
+   |                 help: use the right value directly: `y`
+
+error: aborting due to 3 previous errors
